+++ docs/conf.py
@@ -1,10 +1,9 @@
-# -*- coding: utf-8 -*-
 import os
 import re
 import sys
 
 import sphinx_rtd_theme
-import django_icons
+import django_icons  # noqa
 
 sys.path.insert(0, os.path.abspath("../"))