+++ tests/test_base.py
@@ -1,6 +1,3 @@
-# coding: utf-8
-from __future__ import unicode_literals
-
 from django.test import TestCase
 
 from .test_template_tags import render_template
@@ -13,12 +10,9 @@ class BaseTest(TestCase):
 
     def test_icons(self):
         self.assertEqual(
-            render_template('{% icon "user" size="lg" renderer="BaseRenderer" %}'),
-            '<i class="user"></i>',
+            render_template('{% icon "user" size="lg" renderer="BaseRenderer" %}'), '<i class="user"></i>',
         )
         self.assertEqual(
-            render_template(
-                '{% icon "fas fa-user fa-2x" size="lg" renderer="BaseRenderer" %}'
-            ),
+            render_template('{% icon "fas fa-user fa-2x" size="lg" renderer="BaseRenderer" %}'),
             '<i class="fas fa-user fa-2x"></i>',
         )