+++ tests/test_bootstrap3.py
@@ -1,7 +1,3 @@
-# coding: utf-8
-from __future__ import unicode_literals
-
-from django.template import Template, Context
 from django.test import TestCase
 
 from .test_template_tags import render_template