+++ tests/test_fontawesome.py
@@ -1,6 +1,3 @@
-# coding: utf-8
-from __future__ import unicode_literals
-
 from django.test import TestCase
 
 from .test_template_tags import render_template
@@ -13,8 +10,6 @@ class FontAwesomeTest(TestCase):
 
     def test_icons(self):
         self.assertEqual(
-            render_template(
-                '{% icon "user" size="lg" renderer="FontAwesomeRenderer" %}'
-            ),
+            render_template('{% icon "user" size="lg" renderer="FontAwesomeRenderer" %}'),
             '<i class="fa fa-user fa-lg"></i>',
         )