+++ tests/test_material.py
@@ -1,6 +1,3 @@
-# coding: utf-8
-from __future__ import unicode_literals
-
 from django.test import TestCase
 
 from .test_template_tags import render_template
@@ -13,6 +10,5 @@ class MaterialTest(TestCase):
 
     def test_icons(self):
         self.assertEqual(
-            render_template('{% icon "user" renderer="MaterialRenderer" %}'),
-            '<i class="material-icons">user</i>',
+            render_template('{% icon "user" renderer="MaterialRenderer" %}'), '<i class="material-icons">user</i>',
         )